requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
asyncpg
greenlet
psycopg2-binary